requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.3
djangorestframework==3.14.0
docopt==0.6.2
exceptiongroup==1.1.2
factory-boy==3.2.1
Faker==19.1.0
graphviz==0.20.1
hypothesis==6.81.2
idna==3.4
iniconfig==2.0.0
packaging==23.1
pipreqs==0.4.13
pluggy==1.2.0
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
yarg==0.1.9
